Avoid using .data attribute
Accessing the .data attribute on a compressed image causes the entire
image to be decompressed and read into memory.  That defeats the ability
of astrocut to reduce the IO requirements by reading only the necessary
data.

This change instead uses the .section attribute to access the data,
both when reading the pixels and when checking to see which extensions
have images. The only remaining use of the .data attribute is for access
to the cutout pixels.

Also fixed a bug in the application of the memory_only flag. When true,
it should not write any output files to disk.  However, that was
ignored when single_outfile is False.  This fixes that bug.  It was
already working correctly when single_outfile is True.